// API_Client/client/models/pagination-meta-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My First Nest API
 * The API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface PaginationMetaDto {
    /**
     * ჩანაწერების საერთო რაოდენობა
     */
    'total': number;
    /**
     * მიმდინარე გვერდი
     */
    'page': number;
    /**
     * ჩანაწერები თითო გვერდზე
     */
    'limit': number;
    /**
     * გვერდების საერთო რაოდენობა
     */
    'totalPages': number;
    /**
     * არის თუ არა შემდეგი გვერდი
     */
    'hasNext': boolean;
    /**
     * არის თუ არა წინა გვერდი
     */
    'hasPrevious': boolean;
}